refactor: lecture utils
